# contracts/p5-reporting-decommission/openapi.yaml
openapi: 3.1.0
info:
  title: HRMS Phase 5 – Reporting / Integration, reference-data ownership, Forms + Oracle decommission
  version: 5.0.0-contract
  description: |
    Frozen API surface for Phase 5 (CUTOVER_PLAN.md §9, COMPONENT_MAPPING.md §9 rows
    `PKG_REPORTING` / `PKG_INTEGRATION`, MODERNIZATION_BLUEPRINT.md §8 Reporting row,
    TEST_STRATEGY.md §5 row 5). Written BEFORE any implementation exists; the backend
    node implements exactly this file, the frontend node codes against it.

    Target architecture (non-negotiable, MODERNIZATION_BLUEPRINT.md §9 decisions A/B):
    * Spring Boot (Java 21) on PostgreSQL + React. **No PL/SQL, PL/pgSQL, triggers or
      .pll in the target.** `PKG_REPORTING`, `PKG_INTEGRATION`, the six `VW_*` views and
      `HRMS_ADMIN.fmb` / `HRMS_REPORTS.fmb` are read-only characterization references
      (GOLDEN-ORACLE MODE = OFF: no Oracle runtime exists; expectations are recorded
      fixtures, `legacy_source=recorded`).
    * Three Java modules are introduced (ARCH-01/02, one owner per table):
      - `reporting-module` – **owns no table**. Every report is a read-only Spring Data
        JDBC projection (plain SQL, `WITH RECURSIVE` for the hierarchy) over tables owned
        by the P0–P4 modules, executed on the PostgreSQL read replica when one exists.
        The six `VW_*` views are **not** recreated on PostgreSQL; the report SQL is the
        `tests/reconciliation/pg/*.sql` transcription (Level 3 pins it).
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/reports/employee-directory:
    get:
      tags: [reports]
      operationId: reportEmployeeDirectory
      summary: "`VW_ACTIVE_EMPLOYEES` + `PKG_REPORTING.headcount_report` filters"
      description: |
        One row per `EMPLOYEES.EMPLOYMENT_STATUS='ACTIVE'` employee joined to department,
        job, grade, location and manager. `tenureYears = TRUNC(MONTHS_BETWEEN(asOf,
        hireDate)/12, 1)` (Oracle semantics). `summary.headcountByDepartment` is
        `headcount_report` (`p_as_of_date`, `p_dept_id`, `p_location`): active employees
        with `hireDate <= asOf`, grouped by department, with `avgTenureYears`.
        Sorted by `empId`.
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/DeptId'
        - $ref: '#/components/parameters/LocationCode'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: "Directory rows (`Accept: application/json` → page; `text/csv` → stream)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDirectoryPage'
            text/csv:
              schema: { type: string }
              example: "empId,empNumber,firstName,lastName,fullName,email,phoneWork,hireDate,tenureYears,deptId,deptCode,deptName,costCenter,jobId,jobCode,jobTitle,jobFamily,gradeId,gradeCode,gradeName,locationCode,locationName,city,stateProvince,managerEmpId,managerName\n"
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/employee-directory.csv:
    get:
      tags: [reports]
      operationId: reportEmployeeDirectoryCsv
      summary: CSV alias of `/api/reports/employee-directory` (all rows, no paging)
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/DeptId'
        - $ref: '#/components/parameters/LocationCode'
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reports/org-hierarchy:
    get:
      tags: [reports]
      operationId: reportOrgHierarchy
      summary: "`VW_ORG_HIERARCHY` (`CONNECT BY` → `WITH RECURSIVE`)"
      description: |
        Walks **all** employees from the roots (`managerEmpId IS NULL`) with `WITH
        RECURSIVE` and filters `employmentStatus='ACTIVE'` in the outer select, exactly
        like Oracle applies the view `WHERE` after `CONNECT BY` (active reports of a
        terminated manager keep their `orgLevel` / `orgPath`, `tests/golden/README.md`).
        `orgPath = SYS_CONNECT_BY_PATH(fullName, ' > ')` without the leading separator;
        `isLeaf` = `CONNECT_BY_ISLEAF` (counts terminated children). Cycle protection:
        a manager chain that revisits an `empId` stops at the repeat (legacy would raise
        `ORA-01436`); the row is flagged `cycle=true` – a target-only field, `-20004`
        prevents new cycles at write time (P3). `rootEmpId` restricts to one sub-tree.
        Sorted by `orgPath`.
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - in: query
          name: rootEmpId
          required: false
          description: Sub-tree root (`EMPLOYEES.EMP_ID`); default = all roots.
          schema: { type: integer, format: int64, minimum: 1 }
        - in: query
          name: maxLevel
          required: false
          schema: { type: integer, minimum: 1, maximum: 20 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Hierarchy rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgHierarchyPage'
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EmployeeNotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/org-hierarchy.csv:
    get:
      tags: [reports]
      operationId: reportOrgHierarchyCsv
      summary: CSV alias of `/api/reports/org-hierarchy`
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - in: query
          name: rootEmpId
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - in: query
          name: maxLevel
          required: false
          schema: { type: integer, minimum: 1, maximum: 20 }
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EmployeeNotFound'
  /api/reports/employee-compensation:
    get:
      tags: [reports]
      operationId: reportEmployeeCompensation
      summary: "`VW_EMPLOYEE_COMPENSATION` + `PKG_REPORTING.compensation_summary`"
      description: |
        One row per active employee with the current `SALARY_RECORDS` row
        (`endDate IS NULL`), grade band, `compaRatio = ROUND(baseSalary /
        ((minSalary+maxSalary)/2), 4)` and `yearsInGrade` (`MONTHS_BETWEEN` semantics
        from `effectiveDate`). `summary.byDepartment` reproduces `compensation_summary`
        (`p_dept_id`, `p_grade_id`): `headcount`, `avgSalary`, `minSalary`, `maxSalary`,
        `totalPayroll` per department. **Requires `PAYROLL:VIEW` in addition to
        `REPORTS:VIEW`** – the legacy `HRMS_REPORTS` form hid this report unless
        `has_permission('PAYROLL')`. `baseSalary` is never masked in this report (SEC-05
        applies to SSN/bank data, not salary). Sorted by `empId`.
      x-preauthorize: hasAuthority('REPORTS:VIEW') and hasAuthority('PAYROLL:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/DeptId'
        - in: query
          name: gradeId
          required: false
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Compensation rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeCompensationPage'
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/employee-compensation.csv:
    get:
      tags: [reports]
      operationId: reportEmployeeCompensationCsv
      summary: CSV alias of `/api/reports/employee-compensation`
      x-preauthorize: hasAuthority('REPORTS:VIEW') and hasAuthority('PAYROLL:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/DeptId'
        - in: query
          name: gradeId
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reports/leave-summary:
    get:
      tags: [reports]
      operationId: reportLeaveSummary
      summary: "`VW_LEAVE_SUMMARY` with VAL-05 fixed + `leave_utilization_report`"
      description: |
        One row per `(activeEmployee, leaveType)` balance of `year` (default
        `EXTRACT(YEAR FROM asOf)`). **VAL-05:** `available = openingBalance + accrued -
        used + adjustment - pending` (legacy view omitted `- pending`; accepted final
        difference, README.md). `utilizationPct = ROUND(used*100 / NULLIF(openingBalance
        + accrued, 0), 1)` (null when the denominator is 0, as Oracle). `pending` is
        exposed so the Level 3 pack can assert `legacyAvailable - available == pending`.
        `summary.byLeaveType` reproduces `leave_utilization_report` (`p_year`,
        `p_dept_id`): `employees`, `totalAccrued`, `totalUsed`, `avgUtilizationPct`.
        `EMPLOYEE:VIEW` is not required: the report has no self-scope – `REPORTS:VIEW`
        holders (MANAGER+) see the whole company, as in `HRMS_REPORTS`. Sorted by `empId`,
        `leaveTypeName`.
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/Year'
        - $ref: '#/components/parameters/DeptId'
        - in: query
          name: leaveTypeId
          required: false
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Leave summary rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveSummaryPage'
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/leave-summary.csv:
    get:
      tags: [reports]
      operationId: reportLeaveSummaryCsv
      summary: CSV alias of `/api/reports/leave-summary`
      x-preauthorize: hasAuthority('REPORTS:VIEW')
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/Year'
        - $ref: '#/components/parameters/DeptId'
        - in: query
          name: leaveTypeId
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reports/payroll-latest:
    get:
      tags: [reports]
      operationId: reportPayrollLatest
      summary: "`VW_PAYROLL_LATEST` + `PKG_REPORTING.payroll_summary_report`"
      description: |
        Per-employee aggregates of the **latest `PAYROLL_RUNS` row per period**
        (`MAX(run_id)` per `period_id`, as the view's inline `PAYROLL_RUNS` sub-select),
        default = the latest period with a run in `APPROVED`/`PAID`; `periodId` selects
        another. `STATUS='ERROR'` rows are excluded (P4 contract). `grossPay` /
        `totalTaxes` / `totalDeductions` are positive magnitudes, `netPay = SUM(amount)`
        (P4 sign convention). `summary` is `payroll_summary_report(p_period_id)`:
        `employeeCount`, `totalGross`, `totalTaxes`, `totalDeductions`, `totalNet`,
        `avgNet`. Requires `PAYROLL:VIEW`. Sorted by `empId`.
      x-preauthorize: hasAuthority('REPORTS:VIEW') and hasAuthority('PAYROLL:VIEW')
      parameters:
        - $ref: '#/components/parameters/PeriodId'
        - $ref: '#/components/parameters/DeptId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Payroll rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollLatestPage'
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PeriodNotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/payroll-latest.csv:
    get:
      tags: [reports]
      operationId: reportPayrollLatestCsv
      summary: CSV alias of `/api/reports/payroll-latest`
      x-preauthorize: hasAuthority('REPORTS:VIEW') and hasAuthority('PAYROLL:VIEW')
      parameters:
        - $ref: '#/components/parameters/PeriodId'
        - $ref: '#/components/parameters/DeptId'
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PeriodNotFound'
  /api/reports/pending-approvals:
    get:
      tags: [reports]
      operationId: reportPendingApprovals
      summary: "`VW_PENDING_APPROVALS` (leave requests ∪ performance reviews)"
      description: |
        `UNION ALL` of `LEAVE_REQUESTS.STATUS='PENDING'` (`itemType=LEAVE`) and
        `PERFORMANCE_REVIEWS.STATUS='SUBMITTED'` (`itemType=REVIEW`), each with the
        employee, the approver (`manager_emp_id` / `reviewer_emp_id`) and `daysPending =
        TRUNC(asOf - submittedDate)`. Column names are unified as in the view
        (`itemType, itemId, empId, empName, deptName, approverEmpId, approverName,
        submittedDate, daysPending, detail`). `mine=true` restricts to
        `approverEmpId = jwt.empId` (the manager inbox); without `mine` the caller needs
        `REPORTS:VIEW`, with `mine=true` `isAuthenticated()` suffices. Sorted by
        `submittedDate`, `itemType`, `itemId`.
      x-preauthorize: "hasAuthority('REPORTS:VIEW') or #mine == true"
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - in: query
          name: itemType
          required: false
          schema: { type: string, enum: [LEAVE, REVIEW] }
        - in: query
          name: mine
          required: false
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/DeptId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Pending items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingApprovalPage'
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /api/reports/pending-approvals.csv:
    get:
      tags: [reports]
      operationId: reportPendingApprovalsCsv
      summary: CSV alias of `/api/reports/pending-approvals`
      x-preauthorize: "hasAuthority('REPORTS:VIEW') or #mine == true"
      parameters:
        - $ref: '#/components/parameters/AsOf'
        - in: query
          name: itemType
          required: false
          schema: { type: string, enum: [LEAVE, REVIEW] }
        - in: query
          name: mine
          required: false
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/DeptId'
      responses:
        '200':
          $ref: '#/components/responses/CsvStream'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ---------------------------------------------------------------------------
  # admin – reference data CRUD (admin-module owns the tables from this phase)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: P0 access token; claims `sub` (user-account id, string), `empId`, `authorities[]`.
  parameters:
    AsOf:
      in: query
      name: asOf
      required: false
      description: Replaces `SYSDATE` in the view / ref cursor; default today (server clock, UTC date).
      schema: { type: string, format: date }
    Year:
      in: query
      name: year
      required: false
      description: "`LEAVE_BALANCES.CALENDAR_YEAR`; default `EXTRACT(YEAR FROM asOf)`."
      schema: { type: integer, minimum: 2000, maximum: 2099 }
    DeptId:
      in: query
      name: deptId
      required: false
      description: "`DEPARTMENTS.DEPT_ID` filter (`p_dept_id`)."
      schema: { type: integer, format: int64, minimum: 1 }
    LocationCode:
      in: query
      name: locationCode
      required: false
      description: "`LOCATIONS.LOCATION_CODE` filter (`p_location`)."
      schema: { type: string, maxLength: 10, pattern: '^[A-Z0-9_-]+$' }
    PeriodId:
      in: query
      name: periodId
      required: false
      description: "`PAY_PERIODS.PERIOD_ID`; default = latest period with an APPROVED/PAID run."
      schema: { type: integer, format: int64, minimum: 1 }
...
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      in: query
      name: size
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    CsvStream:
      description: "`text/csv; charset=utf-8`, RFC 4180, header row = JSON property names, all rows, `Content-Disposition: attachment`"
      headers:
        Content-Disposition: { schema: { type: string }, example: 'attachment; filename="leave-summary-2024-06-30.csv"' }
      content:
        text/csv:
          schema: { type: string }
    ValidationFailed:
      description: Bean Validation / unsupported query value (`validation-schema.json` module `p5-reporting-decommission`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: VALIDATION_FAILED, message: asOf must be a date (YYYY-MM-DD), field: asOf, traceId: 4f3c… }
    Unauthorized:
      description: Missing / invalid / revoked token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: TOKEN_INVALID, message: Access token is missing or invalid, traceId: 4f3c… }
    Forbidden:
      description: "`@PreAuthorize` denied (`REPORTS:VIEW`, `PAYROLL:VIEW`, `ADMIN:VIEW`, `ADMIN:EDIT`, `LEAVE:ADMIN`, `PAYROLL:APPROVE`, `PAYROLL:EDIT`, `EMPLOYEE:VIEW`)"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: FORBIDDEN, message: Access denied, traceId: 4f3c… }
    NotAcceptable:
      description: "`Accept` is neither `application/json` nor `text/csv`"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: NOT_ACCEPTABLE, message: Supported representations are application/json and text/csv, traceId: 4f3c… }
    EmployeeNotFound:
      description: "`-20001`: `rootEmpId` / `managerEmpId` / `emp_number` matches no active employee"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: '-20001', message: 'Employee not found or not active: 999', field: managerEmpId, traceId: 4f3c… }
    PeriodNotFound:
      description: "`PERIOD_NOT_FOUND`: `periodId` matches no `PAY_PERIODS` row (P4 framework code)"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: PERIOD_NOT_FOUND, message: 'Pay period not found: 202407', field: periodId, traceId: 4f3c… }
...
  schemas:
...
    Years:
      type: string
      pattern: '^-?[0-9]+\.[0-9]$'
      description: "`TRUNC(MONTHS_BETWEEN(...)/12, 1)` – one decimal, Oracle semantics."
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    EmployeeDirectoryRow:
      type: object
      description: "`VW_ACTIVE_EMPLOYEES` row; CSV column order = property order below."
      required: [empId, empNumber, firstName, lastName, fullName, hireDate, tenureYears, deptId, deptCode, deptName, jobId, jobCode, jobTitle, gradeId, gradeCode]
      properties:
        empId: { type: integer, format: int64 }
        empNumber: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        fullName: { type: string, description: "`FIRST_NAME || ' ' || LAST_NAME`" }
        email: { type: [string, 'null'] }
        phoneWork: { type: [string, 'null'] }
        hireDate: { type: string, format: date }
        tenureYears: { $ref: '#/components/schemas/Years' }
        deptId: { type: integer, format: int64 }
        deptCode: { type: string }
        deptName: { type: string }
        costCenter: { type: [string, 'null'] }
        jobId: { type: integer, format: int64 }
        jobCode: { type: string }
        jobTitle: { type: string }
        jobFamily: { type: [string, 'null'] }
        gradeId: { type: integer }
        gradeCode: { type: string }
        gradeName: { type: string }
        locationCode: { type: [string, 'null'] }
        locationName: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        stateProvince: { type: [string, 'null'] }
        managerEmpId: { type: [integer, 'null'], format: int64 }
        managerName: { type: [string, 'null'] }
...
    EmployeeDirectoryPage:
      type: object
      required: [asOf, content, page, summary]
      properties:
        asOf: { type: string, format: date }
        content:
          type: array
          items: { $ref: '#/components/schemas/EmployeeDirectoryRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
        summary:
          type: object
          required: [totalHeadcount, headcountByDepartment]
          properties:
            totalHeadcount: { type: integer, minimum: 0 }
            headcountByDepartment:
              type: array
              items: { $ref: '#/components/schemas/HeadcountByDepartment' }
...
    OrgHierarchyPage:
      type: object
      required: [asOf, content, page]
      properties:
        asOf: { type: string, format: date }
        content:
          type: array
          items: { $ref: '#/components/schemas/OrgHierarchyRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    EmployeeCompensationPage:
      type: object
      required: [asOf, content, page, summary]
      properties:
        asOf: { type: string, format: date }
        content:
          type: array
          items: { $ref: '#/components/schemas/EmployeeCompensationRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
        summary:
          type: object
          required: [byDepartment]
          properties:
            byDepartment:
              type: array
              items: { $ref: '#/components/schemas/CompensationByDepartment' }
...
    LeaveSummaryPage:
      type: object
      required: [asOf, year, content, page, summary]
      properties:
        asOf: { type: string, format: date }
        year: { type: integer }
        content:
          type: array
          items: { $ref: '#/components/schemas/LeaveSummaryRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
        summary:
          type: object
          required: [byLeaveType]
          properties:
            byLeaveType:
              type: array
              items: { $ref: '#/components/schemas/LeaveUtilizationByType' }
...
    PayrollLatestPage:
      type: object
      required: [content, page, summary]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/PayrollLatestRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
        summary: { $ref: '#/components/schemas/PayrollSummary' }
...
    PendingApprovalPage:
      type: object
      required: [asOf, content, page, summary]
      properties:
        asOf: { type: string, format: date }
        content:
          type: array
          items: { $ref: '#/components/schemas/PendingApprovalRow' }
        page: { $ref: '#/components/schemas/PageMeta' }
        summary:
          type: object
          required: [leave, review]
          properties:
            leave: { type: integer, minimum: 0 }
            review: { type: integer, minimum: 0 }

    # --- admin reference data ----------------------------------------------------
...